Move plugin into its own module
